requirements.txt
flask
requests
markdown
Pillow
pytesseract
